refactor rest package
