+++ code/go/0chain.net/smartcontract/minersc/benchmark_rest_tests.go
@@ -9,7 +9,7 @@ import (
 func BenchmarkRestTests(
 	data benchmark.BenchData, _ benchmark.SignatureScheme,
 ) benchmark.TestSuite {
-	rh := rest.NewTestRestHandler()
+	rh := rest.NewRestHandler(&rest.TestQueryChainer{})
 	mrh := NewMinerRestHandler(rh)
 	return benchmark.GetRestTests(
 		[]benchmark.TestParameters{
