+++ code/go/0chain.net/smartcontract/vestingsc/benchmark_rest_tests.go
@@ -10,7 +10,7 @@ const owner = "1746b06bb09f55ee01b33b5e2e055d6cc7a900cb57c0a3a5eaabb8a0e7745802"
 func BenchmarkRestTests(
 	data benchmark.BenchData, _ benchmark.SignatureScheme,
 ) benchmark.TestSuite {
-	rh := rest.NewTestRestHandler()
+	rh := rest.NewRestHandler(&rest.TestQueryChainer{})
 	vrh := NewVestingRestHandler(rh)
 	return benchmark.GetRestTests(
 		[]benchmark.TestParameters{
